Organize bean packages
