+++ dev/com.ibm.ws.jsf.2.3_fat/test-applications/CDICommon/src/com/ibm/ws/jsf23/fat/cdi/common/beans/jsf23/NavigationHandlerBean.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.jsf23.fat.cdi.common.beans;
+package com.ibm.ws.jsf23.fat.cdi.common.beans.jsf23;
 
 import java.io.Serializable;
